RFC: Make the `axis_buffer` kwarg work again
The code ast example in the README does not look right because the `axis_buffer` kwarg is not working. Unfortunately, I had to add NaNMath as a dependency for this to work. I will leave this PR open for some time so that people can comment on the added dependency.