Axis buffer (#62)
* Fix spectral method

Without this PR, the added test would fail.

* Use the spectral method from NetworkLayouts

I have deleted the test, since NetworkLayouts has it's own tests. We will be able to test this method and others when we have generated documentation.

* RFC: Make the `axis_buffer` kwarg work again

The code ast example in the README does not look right because the `axis_buffer` kwarg is not working. Unfortunately, I had to add NaNMath as a dependency for this to work. I will leave this PR open for some time so that people can comment on the added dependency.